prompttech/requirements.txt
django-admin-interface
django-colorfield
django-tinymce
django-import-export
django-extensions
python-decouple
pillow